j'ai regle ce que j'avais a regler, leaks a revoir et un peu de nettoyage

+++ src/ft_heredoc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_heredoc.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: agouin <agouin@42.fr>                      +#+  +:+       +#+        */
+/*   By: skuor <skuor@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/06 12:39:18 by agouin            #+#    #+#             */
-/*   Updated: 2025/09/19 16:32:09 by agouin           ###   ########.fr       */
+/*   Updated: 2025/10/05 16:20:42 by skuor            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,8 +85,10 @@ int	ft_setup_heredoc(t_cmd *commande)
 	{
 		waitpid(pid, &status, 0);
 		sigaction(SIGINT, &old_signale, NULL);
+		int exit_code = WEXITSTATUS(status);
+   		if (exit_code == 130)
+			return (-1);
 		fd = open(".files", O_RDONLY);
-		//close(fd);
 		unlink(".files");
 		return (fd);
 	}
